+++ README.md
@@ -283,7 +283,7 @@ var colorChangeTool: Tool {
 
 *CrashDetectionTool* sets up listeners on system exceptions which will be logged into a local json file with the stack trace. The tool is available on iOS.
 
-#### Database Import/Export tool
+#### Database tool
 
 *DatabaseImportExportTool* gives the ability to import or export the database file which is located in the documents folder of the app. The tool is available on iOS and MacOS.
