+++ Sentinel/Classes/DatabaseImportExport/DatabaseTool.swift
@@ -1,5 +1,5 @@
 //
-//  DatabaseImportExportTool.swift
+//  DatabaseTool.swift
 //  Sentinel
 //
 //  Created by Zvonimir Medak on 12.03.2025..
@@ -9,7 +9,7 @@ import SwiftUI
 import UniformTypeIdentifiers
 
 /// Tool which gives the ability to import or export the database file
-public struct DatabaseImportExportTool: Tool {
+public struct DatabaseTool: Tool {
 
     // MARK: - Public properties
 
@@ -27,7 +27,7 @@ public struct DatabaseImportExportTool: Tool {
     ///   - databaseFilePath: Relative path of the database file from the Documents directory
     ///   - allowedTypes: Types of the files which can be imported, should be the type of the database file
     public init(
-        name: String = "Database Import/Export Tool",
+        name: String = "Database Tool",
         databaseFilePath: String,
         allowedTypes: [UTType]
     ) {
@@ -39,12 +39,12 @@ public struct DatabaseImportExportTool: Tool {
 
 // MARK: - UI
 
-public extension DatabaseImportExportTool {
+public extension DatabaseTool {
 
     #if os(macOS)
     func createContent(selection: Binding<String?>) -> any View {
-        DatabaseImportExportView(
-            viewModel: DatabaseImportExportViewModel(
+        DatabaseView(
+            viewModel: DatabaseViewModel(
                 databaseFilePath: databaseFilePath,
                 allowedTypes: allowedTypes
             )
@@ -52,8 +52,8 @@ public extension DatabaseImportExportTool {
     }
     #else
     var content: any View {
-        DatabaseImportExportView(
-            viewModel: DatabaseImportExportViewModel(
+        DatabaseView(
+            viewModel: DatabaseViewModel(
                 databaseFilePath: databaseFilePath,
                 allowedTypes: allowedTypes
             )
